Replaced long with proper 64bit integers Long is only 32bit on windows and thus breaks all calculations it was specifically used in Also improved shortDiff with new diffUnsigned - this allows for signed differences smaller than full unsigned integers without unexpected overflows Also expanded useage of cross-platform printf integer notations - Need to use PRIu64/PRId64 instead of ld/lu, again because those expect 32bit integers on windows

+++ server/source/comm/streaming.hpp
@@ -101,7 +101,7 @@ struct SyncGroup
 	StatDistf SOFSwitch;
 
 	// Stats
-	long long frameCount = 0;
+	uint64_t frameCount = 0;
 	int frameProcessedCount; // Frames that had any data to be processed at all
 	int packetMissingCount; // Frames that had missing packets from cameras
 	int packetErroneousCount; // Frames that had erroneous packets from cameras
